Updat dog breeds code - add app

+++ code/dl/cnn/dog_breeds/data_downloader.py
@@ -8,37 +8,63 @@ import shutil
 from zipfile import ZipFile
 import pandas as pd
 from sklearn.model_selection import train_test_split
+from shutil import rmtree
 
 
 KAGGLE_DATA_URL = "https://storage.googleapis.com/teamlab-data/dog-breed-identification.zip"
 STANFORD_DATA_URL = "https://storage.googleapis.com/teamlab-data/stanford-dogs-dataset.zip"
 DATA_DIR = "data"
-#FLAGS = flags.FLAGS
-#flags.DEFINE_string("name", None, "Your name.")
-#flags.DEFINE_boolean("stanford_dataset", False,
-                     #}"Whether to use Stanford dataset.")
+FLAGS = flags.FLAGS
+flags.DEFINE_boolean("stanford_dataset", False,
+                     "Whether to use Stanford dataset.")
+flags.DEFINE_boolean("reset", False,
+                     "Whether to remove all downloaded dataset")
+flags.DEFINE_float("test_ratio", 0.2,
+                     "Whether to remove all downloaded dataset")
 
 # Required flag.
 #flags.mark_flag_as_required("name")
 
-def kaggle_data_downlad():
-  download_path = os.path.join(DATA_DIR, "dog-breed-identification.zip")
-  wget.download(KAGGLE_DATA_URL, download_path)
+def _data_downlad(dataset_name):
+  if dataset_name == "KAGGLE":
+    download_path = os.path.join(DATA_DIR, "dog-breed-identification.zip")
+    wget.download(KAGGLE_DATA_URL, download_path)
+  if dataset_name == "STANFORD":
+    download_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip")
+    wget.download(STANFORD_DATA_URL, download_path)
 
-def stanford_data_downlad():
-  download_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip")
-  wget.download(STANFORD_DATA_URL, download_path)
+def _write_current_download_configuration():
+  import json 
 
+  conf = dict()
+  if FLAGS.stanford_dataset:
+    conf["dataset"] = "STANFORD"
+  else:
+    conf["dataset"] = "KAGGLE_ONLY"
+  
+  conf["test_ratio"] = FLAGS.test_ratio
+
+  with open('download_conf.json', 'w') as outfile:
+    json.dump(conf, outfile)
+  
 def main(args):
+  _write_current_download_configuration()
+
+  if FLAGS.reset == True:
+    print("Now, deleting current all downloaded dataset")
+    rmtree(DATA_DIR)
+    print("Done")
+
   if os.path.exists(DATA_DIR) is not True:
     os.mkdir(DATA_DIR)
+    print("Make 'data' directory ")
   
   # Kaggle Dataset Download
   kaggle_data_path = os.path.join(DATA_DIR, "dog-breed-identification.zip") 
   if os.path.exists(kaggle_data_path) is not True:
     print("kaggle dog breed dataset download")
-    kaggle_data_downlad()
-  print("Kaggle dog breed dataset download is done")
+    _data_downlad("KAGGLE")
+  print("\nKaggle dog breed dataset download is done")
   
 
 
@@ -58,8 +84,10 @@ def main(args):
   x = label_df.id + ".jpg"
   y = label_df.breed
   
+  print ("Train / Validation Split: Split ratio - {0}".format(
+      FLAGS.test_ratio))
   x_train_files, x_val_files, y_train_labels, y_val_labels = train_test_split(
-    x, y, test_size = 0.2, stratify = y)
+    x, y, test_size = FLAGS.test_ratio, stratify = y)
   
   VAL_DIR = "val"
   TRAIN_DIR = "train"
@@ -84,26 +112,32 @@ def main(args):
     shutil.move(source_path, target_path)
 
   # Stanford Dataset Download
-  stanford_data_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip") 
-  if os.path.exists(stanford_data_path) is not True:
-    print("Stanford dataset download")
-    stanford_data_downlad()
-  print("Stanford dog breed dataset download is done")
-  
-  stanford_dataset_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip")
-  with ZipFile(stanford_dataset_path, 'r') as zipObj:
-     zipObj.extractall(DATA_DIR)
-  stanford_image_path = os.path.join(DATA_DIR, "images", "Images")
-  for breed_name in os.listdir(stanford_image_path):
-    label = breed_name[10:].lower()
-    breed_path = os.path.join(stanford_image_path, breed_name)
-    for filename in os.listdir(breed_path):
-      source_path = os.path.join(breed_path, filename)
-      target_path = os.path.join(DATA_DIR, TRAIN_DIR, label, filename) 
-      shutil.move(source_path, target_path)
-  shutil.rmtree(os.path.join(DATA_DIR, "images"))
+  if FLAGS.stanford_dataset == True:
+    stanford_data_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip") 
+    if os.path.exists(stanford_data_path) is not True:
+      print("Stanford dataset download")
+      _data_downlad("STANFORD")
+    print("\nStanford dog breed dataset download is done")
+    
+    stanford_dataset_path = os.path.join(DATA_DIR, "stanford-dogs-dataset.zip")
+    with ZipFile(stanford_dataset_path, 'r') as zipObj:
+      zipObj.extractall(DATA_DIR)
+    stanford_image_path = os.path.join(DATA_DIR, "images", "Images")
+    for breed_name in os.listdir(stanford_image_path):
+      label = breed_name[10:].lower()
+      breed_path = os.path.join(stanford_image_path, breed_name)
+      for filename in os.listdir(breed_path):
+        source_path = os.path.join(breed_path, filename)
+        target_path = os.path.join(DATA_DIR, TRAIN_DIR, label, filename) 
+        shutil.move(source_path, target_path)
+    shutil.rmtree(os.path.join(DATA_DIR, "images"))
+
+    file_count = sum([len(files) for r, d, files in os.walk(train_path)])
+    print("Number of train files : {0}".format(file_count))
+
+    file_count = sum([len(files) for r, d, files in os.walk(val_path)])
+    print("Number of validation files : {0}".format(file_count))
 
 if __name__ == '__main__':
   app.run(main)
 
-  #wget.download(url, 'cat4.jpg')
